Conduct tracking for artifact use.
This is broken down into three categories, tracked separately:
 * Equipping (wearing or wielding) an artifact.

 * Applying or invoking an artifact.
   (For some, e.g., eyewear, applying may also count as equipping.)

 * Carrying an artifact.  This only counts if you move while the
   artifact is in open inventory, so you can e.g. pick up your quest
   artifact and immediately bag it, then later take it out and show
   it to your quest leader and immediately drop or bag it again,
   without breaking the conduct.

If you keep all three conducts, the game simply says that you did not
"use" any artifacts.

+++ libnethack/include/extern.h
@@ -1811,7 +1811,7 @@ extern int wiz_timeout_queue(const struct nh_cmd_arg *);
 /* ### topten.c ### */
 
 extern void munge_xlstring(char *, const char *, int);
-extern unsigned long encode_conduct(void);
+extern unsigned long long encode_conduct(void);
 extern void update_topten(int how, const char *killer,
                           unsigned long carried, const char *dumpname);
 extern struct obj *tt_oname(struct obj *);
